requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Celery for background tasks
celery[redis]==5.3.4
flower==2.0.1

# Database and storage
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# AI/ML Dependencies
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
xformers>=0.0.22

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# Real-ESRGAN for upscaling
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8
realesrgan>=0.3.0

# SAM2 (Segment Anything Model 2)
# Note: May need to install from source or specific repo
# git+https://github.com/facebookresearch/segment-anything-2.git

# Utilities
pydantic>=2.4.0
python-dotenv>=1.0.0
loguru>=0.7.2
tqdm>=4.66.0
requests>=2.31.0
psutil>=5.9.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Optional: For face enhancement
# gfpgan
# codeformer

# Optional: For better performance
# onnxruntime-gpu  # if using ONNX models